Relocated API + added SpawnObject method
