+++ NewHorizons/Main.cs
@@ -427,72 +427,4 @@ namespace NewHorizons
         }
         #endregion Change star system
     }
-
-    #region API
-    public class NewHorizonsApi
-    {
-        [Obsolete("Create(Dictionary<string, object> config) is deprecated, please use Create(Dictionary<string, object> config, IModBehaviour mod) instead")]
-        public void Create(Dictionary<string, object> config)
-        {
-            Create(config, null);
-        }
-
-        public void Create(Dictionary<string, object> config, IModBehaviour mod)
-        {
-            Logger.Log("Recieved API request to create planet " + (string)config["Name"], Logger.LogType.Log);
-            var planetConfig = new PlanetConfig(config);
-
-            var body = new NewHorizonsBody(planetConfig, mod ?? Main.Instance);
-
-            if (!Main.BodyDict.ContainsKey(body.Config.StarSystem)) Main.BodyDict.Add(body.Config.StarSystem, new List<NewHorizonsBody>());
-            Main.BodyDict[body.Config.StarSystem].Add(body);
-        }
-
-        public void LoadConfigs(IModBehaviour mod)
-        {
-            Main.Instance.LoadConfigs(mod);
-        }
-
-        public GameObject GetPlanet(string name)
-        {
-            return Main.BodyDict.Values.SelectMany(x => x)?.ToList()?.FirstOrDefault(x => x.Config.Name == name)?.Object;
-        }
-
-        public string GetCurrentStarSystem()
-        {
-            return Main.Instance.CurrentStarSystem;
-        }
-
-        public UnityEvent<string> GetChangeStarSystemEvent()
-        {
-            return Main.Instance.OnChangeStarSystem;
-        }
-
-        public UnityEvent<string> GetStarSystemLoadedEvent()
-        {
-            return Main.Instance.OnStarSystemLoaded;
-        }
-
-        public bool ChangeCurrentStarSystem(string name)
-        {
-            if (!Main.SystemDict.ContainsKey(name)) return false;
-
-            Main.Instance.ChangeCurrentStarSystem(name);
-            return true;
-        }
-
-        public string[] GetInstalledAddons()
-        {
-            try
-            {
-                return Main.MountedAddons.Select(x => x?.ModHelper?.Manifest?.UniqueName).ToArray();
-            }
-            catch (Exception ex)
-            {
-                Logger.LogError($"Couldn't get installed addons {ex.Message}, {ex.StackTrace}");
-                return new string[] { };
-            }
-        }
-    }
-    #endregion API
 }
